refactor: address architect review findings
- Split test-harness.ts into pg-lifecycle.ts (Postgres infra) and
  bench-runner.ts (test assertion logic)
- Replace inline worker loops in copy-threshold.ts and copy-crossover.ts
  with scopedBatchWriter
- Fix inline import() type in bulk-import.ts to use direct import
- Delete unused barrel file scenarios/index.ts
