+++ packages/event-store-bench/src/scenarios/copy-threshold.ts
@@ -1,7 +1,6 @@
-import { Tags } from "@dcb-es/event-store"
 import { ThresholdFactory, ScenarioResult } from "../harness/types"
 import { buildScenarioResult } from "../harness/stats"
-import { emptyResult, spawnWorkers } from "../harness/workers"
+import { scopedBatchWriter, spawnWorkers } from "../harness/workers"
 import { ThresholdBenchScenario, StressTestResult } from "./types"
 
 interface Config {
@@ -18,33 +17,13 @@ async function run(factory: ThresholdFactory, config: Config): Promise<StressTes
         const store = await factory(`copy-thresh-${threshold}-${Date.now()}`, threshold)
         const endTime = Date.now() + config.durationPerTierMs
 
-        const workers = await spawnWorkers(config.workerCount, async (id) => {
-            const result = emptyResult(id, "write")
-            const scopeTag = `W${id}`
-            let iteration = 0
-
-            while (Date.now() < endTime) {
-                const events = Array.from({ length: config.batchSize }, (_, i) => ({
-                    type: "ThresholdEvent",
-                    tags: Tags.fromObj({ scope: scopeTag }),
-                    data: { seq: iteration * config.batchSize + i },
-                    metadata: {},
-                }))
-
-                const opStart = Date.now()
-                try {
-                    await store.append({ events })
-                    result.latencies.push(Date.now() - opStart)
-                    result.operations++
-                    result.events += config.batchSize
-                } catch {
-                    result.errors++
-                }
-                iteration++
-            }
-
-            return result
-        })
+        const workers = await spawnWorkers(config.workerCount, (id) =>
+            scopedBatchWriter(store, id, endTime, {
+                batchSize: config.batchSize,
+                eventType: "ThresholdEvent",
+                conditional: false,
+            }),
+        )
 
         const route = config.batchSize <= threshold ? "function" : "COPY"
         scenarios.push(buildScenarioResult(
