--- packages/event-store-bench/src/scenarios/index.ts
@@ -1,18 +1,0 @@
-export { throughputScaling } from "./throughput-scaling"
-export { bulkImport } from "./bulk-import"
-export { contention } from "./contention"
-export { mixedWorkload } from "./mixed-workload"
-export { rawThroughput } from "./raw-throughput"
-export { parallelImport } from "./parallel-import"
-export { consistencyOracle } from "./consistency-oracle"
-export { overlapConsistency } from "./overlap-consistency"
-export { esbCompat } from "./esb-compat"
-export { meterUpsert } from "./meter-upsert"
-export { batchSweepUnconditional, batchSweepConditional } from "./batch-sweep"
-export { degradation } from "./degradation"
-export { batchCommands } from "./batch-commands"
-
-export { copyThreshold } from "./copy-threshold"
-export { copyCrossover } from "./copy-crossover"
-
-export type { StressTestResult, VerificationResult, BenchScenario, ThresholdBenchScenario } from "./types"
